--- src/test/java/com/shapesecurity/shift/es2016/parser/Test262Test.java
@@ -1,106 +1,0 @@
-package com.shapesecurity.shift.es2016.parser;
-
-import com.shapesecurity.shift.es2016.ast.Module;
-import com.shapesecurity.shift.es2016.ast.Script;
-import com.shapesecurity.shift.es2016.serialization.Deserializer;
-import org.json.JSONException;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
-
-import java.io.File;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import static org.junit.Assert.assertEquals;
-
-@RunWith(Parameterized.class)
-public class Test262Test {
-	static final String testsDir = "test/test262-parser-tests/";
-	static final String expectationsDir = "test/shift-parser-expectations/expectations/";
-
-	private static final Set<String> xfail = new HashSet<>(Arrays.asList(
-			"05b849122b429743.js" // Java's unicode support appears to be out of date
-	));
-	// TODO have some test to ensure all xfail tests actually exist
-
-	private static void checkPass(String name) throws JsError, IOException, ClassNotFoundException, NoSuchMethodException, InstantiationException, JSONException, IllegalAccessException {
-		// TODO check locations (requires deserializer subclass)
-
-		String src = new String(Files.readAllBytes(Paths.get(testsDir, "pass", name)), StandardCharsets.UTF_8);
-		String expectedJSON = new String(Files.readAllBytes(Paths.get(expectationsDir, name + "-tree.json")), StandardCharsets.UTF_8);
-
-		if (name.endsWith(".module.js")) {
-			Module actual = Parser.parseModule(src);
-			Module expected = (Module) Deserializer.deserialize(expectedJSON);
-			assertEquals(expected, actual);
-		} else {
-			Script actual = Parser.parseScript(src);
-			Script expected = (Script) Deserializer.deserialize(expectedJSON);
-			assertEquals(expected, actual);
-		}
-	}
-
-	@Parameterized.Parameters(name = "{1} : {0}")
-	public static Collection<Object[]> data() {
-		File[] passFiles = (new File(testsDir + "pass")).listFiles();
-		Stream<Object[]> pass = Arrays.stream(passFiles)
-				.map(f -> new Object[]{ f.getName(), TestType.PASS });
-
-		File[] earlyFiles = (new File(testsDir + "early")).listFiles();
-		Stream<Object[]> early = Arrays.stream(earlyFiles)
-				.map(f -> new Object[]{ f.getName(), TestType.EARLY });
-
-		File[] failFiles = (new File(testsDir + "early")).listFiles();
-		Stream<Object[]> fail = Arrays.stream(failFiles)
-				.map(f -> new Object[]{ f.getName(), TestType.FAIL });
-
-		return Stream.of(pass, early, fail)
-				.flatMap(x -> x)
-				.collect(Collectors.toList()); // TODO consider just splitting this into multiple test classes
-	}
-
-	@Parameterized.Parameter
-	public String name;
-
-	@Parameterized.Parameter(1)
-	public TestType type;
-
-	@Test
-	public void test() throws Exception {
-		boolean passed = false;
-		try {
-			switch (this.type) {
-				case PASS:
-					checkPass(this.name);
-					break;
-				case EARLY:
-					// TODO
-				case FAIL:
-					// TODO
-			}
-			passed = true;
-		} catch (Exception e) {
-			if (!xfail.contains(name)) {
-				throw e;
-			} // else swallow the exception and thus pass
-		}
-		if (passed && xfail.contains(name)) {
-			throw new RuntimeException("Test marked as xfail, but passed");
-		}
-	}
-
-	public enum TestType {
-		PASS,
-		EARLY,
-		FAIL
-	}
-}
